backend/requirements.txt
tflite-runtime>=2.12.0
flask>=2.0.0
flask-cors>=3.0.0
pillow>=9.0.0
numpy>=1.19.0